fix: export a value of 0 for all counter metrics at startup
At the moment, this is done manually, which is a solution I'm not super
happy with. I'd like to abstract over this at some point.

The intention behind the change is that the Rust
metrics-exporter-prometheus crate apparently *doesn't* do this, unlike
other prometheus exporter implementations - and it means that, for
example, the increase() calculated for the first time the counter is
incremented after startup is going to come out as 0.